% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.graphs.R
\name{plotx}
\alias{plotx}
\title{Displays the map of the explanatory variables}
\usage{
plsda.plotx(objetPLSDA)
}
\arguments{
\item{objetPLSDA, }{an object returned by the function plsda.fit.}
}
\value{
The function returns the map of the explanatory variables.
}
\description{
Displays the map of the explanatory variables
}
