% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.dummies.R
\name{plsda.dummies}
\alias{plsda.dummies}
\title{This function transforms a vector of \code{p} factors into a matrix of \code{p} dummies variables.}
\usage{
plsda.dummies(X, name)
}
\arguments{
\item{X}{the vector of factors to transform.}

\item{name}{a vector containing the original variables names that will be the \code{colnames} of the dummy matrix.}
}
\value{
The function returns a dummy matrix with \code{p} columns, named with the vector "\code{name}".
}
\description{
This function transforms a vector of \code{p} factors into a matrix of \code{p} dummies variables.
}
\examples{
dummies.t1<-plsda.dummies(iris$Species)
dummies.t2<-plsda.dummies(iris$Species,"Species")
}
