% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.graphs.R
\name{plsda.Imap}
\alias{plsda.Imap}
\title{Displays the map of the individuals}
\usage{
plsda.Imap(objetPLSDA)
}
\arguments{
\item{objetPLSDA, }{an object returned by the function plsda.fit.}
}
\value{
The function returns the map of the individuals.
}
\description{
This function displays a plot.
}
