% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{plsda.predict}
\alias{plsda.predict}
\title{Predict function
It is used either to predict the modality of a target variable depending on the explanatory variables.}
\usage{
plsda.predict(objetPLSDA, newdata, type = "class")
}
\arguments{
\item{objetPLSDA, }{an object belonging to the class PLSDA, an object which was returned by the fit function. 
The model obtained with our train dataset}

\item{newdata, }{the test dataset, with a number of observation and values for each explanatory variable.}

\item{type, }{the type of prediction we want. If the type is posterior, the function returns 
the probability of belonging to each modality.
If the type is class, the function returns the modality predicted for each observation.}
}
\value{
A column vector containing the predicted values for the target variable, according to the test dataset.
}
\description{
Predict function
It is used either to predict the modality of a target variable depending on the explanatory variables.
}
\examples{
pred <- plsda.predict(fit1,iris$test, type='class')
pred <- plsda.predict(fit1,iris$test, type='posterior')

}
