% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.scale.R
\name{plsda.scale}
\alias{plsda.scale}
\title{this function allows to center and reduce the matrix or the numerical vector X of the predictive variables.}
\usage{
plsda.scale(X,reduce = F)
}
\arguments{
\item{X}{The matrix that must be centered and standardized if \code{reduce = T}.}

\item{reduce}{is an optional parameters which if \code{TRUE} standardizes the data frame or numeric vector X.}
}
\value{
The function returns a list of at least the following components :

\code{New} is the resulting centered (and standardized if \code{reduce = T}) matrix.
\cr
\code{means} is the average of the initial matrix or numeric vector X.
\cr
\code{vars} is the variance of the initial matrix or numeric vector X.
}
\description{
this function allows to center and reduce the matrix or the numerical vector X of the predictive variables.
}
\examples{
scale.t1<-plsda.scale(iris[,-5])
scale.t2<-plsda.scale(iris[,-5],reduce=T)
}
