% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_graphs.R
\name{plsda.scree}
\alias{plsda.scree}
\title{Displays the plot scree of the components importance}
\usage{
plsda.scree(PLSDAObject)
}
\arguments{
\item{PLSDAObject, }{an object returned by the function plsda.fit, the model the user is using.}
}
\value{
The function returns a screeplot, with the components on the abscissa axis and their eigen values on the ordinate axis.
}
\description{
This function displays a screeplot on which we can see the importance of each component by calculating their eigenvalues.
}
