% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.graphs.R
\name{plsda.vmap}
\alias{plsda.vmap}
\title{Displays the map of the explanatory variables}
\usage{
plsda.vmap(objetPLSDA)
}
\arguments{
\item{PLSDAObject, }{an object returned by the function plsda.fit.}
}
\value{
The function returns a screeplot.
}
\description{
This function displays a screeplot on which we can see the importance of each explanatory variable each explanatory variable by multiplying the square root of their eigenvalues with their loadings (Xloadings).
}
