% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.fit.R
\name{plsda.fit}
\alias{plsda.fit}
\title{PLS-DA
{plsda.fit} is used to fit a PLS-DA regression model.}
\usage{
plsda.fit(formula, data, ncomp = 2)
}
\arguments{
\item{formula, }{an object of class formula: a symbolic description of the model
to be fitted. The details of model specification are given under ‘Details'.}

\item{data, }{the dataframe containing the variables in the model.}

\item{ncomp, }{the number of components extracted in NIPALS algorithm, an integer
comprised between 1 and the number of explanatory variables.}
}
\value{
An object of class 'PLSDA' containing the following components :

{X} the original dataset containing the predictors.

{Y} the variable(s) to predict.

{level} the terms of the variables to predict.

{Xmeans} the means of each predictor.

{pls.coef}the loadings of the predictors, calculated by NIPALS algorithm in plsda.pls function.

{ncomp} the number of components used in plsda.pls.

{locoef} the coefficients of the logistic regression applied on the loadings of the predictors
to have the final coefficients of PLS-DA.

{plsda.coef} the final coefficients of the plsda, which are logit functions.
}
\description{
PLS-DA
{plsda.fit} is used to fit a PLS-DA regression model.
}
\examples{
#ncomp is specified
fit1<-plsda.fit(Species~.,iris,ncomp=4)
}
