% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_graphs.R
\name{plsda.Imap}
\alias{plsda.Imap}
\title{Displays the map of the individuals}
\usage{
plsda.Imap(PLSDAObject)
}
\arguments{
\item{PLSDAObject, }{an object returned by the function plsda.fit, the model the user is using.}
}
\value{
The function returns a screeplot, with the components on the abscissa axis and their eigen values on the ordinate axis.
}
\description{
This function displays a.
}
