% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_graphs.R
\name{plsda.vmap}
\alias{plsda.vmap}
\title{Displays the map of the explicative variables}
\usage{
plsda.vmap(PLSDAObject)
}
\arguments{
\item{PLSDAObject, }{an object returned by the function plsda.fit, the model the user is using.}
}
\value{
The function returns a .
}
\description{
This function displays a screeplot on which we can see the importance of each component by calculating their eigenvalues.
}
